Fix error/warning handling with multiple files

+++ oc/compiler/symtab/symtab.c
@@ -860,7 +860,7 @@ void print_type_name(symtab_type_record_t* record){
  * Crawl the symtab and check for any unused functions. We generate some hopefully helpful
  * warnings here for the user
  */
-void check_for_unused_functions(function_symtab_t* symtab, u_int16_t* num_warnings){
+void check_for_unused_functions(function_symtab_t* symtab, u_int32_t* num_warnings){
 	//For any/all error printing
 	char info[1000];
 	//For temporary holding
@@ -911,7 +911,7 @@ void check_for_unused_functions(function_symtab_t* symtab, u_int16_t* num_warnin
  * If a variable is declared as "mut"(mutable) but is never assigned to throughout it's
  * entire lifetime, that mut keyword is not needed
  */
-void check_for_var_errors(variable_symtab_t* symtab, u_int16_t* num_warnings){
+void check_for_var_errors(variable_symtab_t* symtab, u_int32_t* num_warnings){
 	//For any/all error printing
 	char info[1000];
 	//For record holding
